@@ -7,20 +7,21 @@
     "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
     "start": "nx serve",
     "build": "nx build",
+    "lint": "nx workspace-lint && ng lint",
     "test": "nx test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.2.0",
-    "@angular/cdk": "^13.2.6",
-    "@angular/common": "~13.2.0",
-    "@angular/compiler": "~13.2.0",
-    "@angular/core": "~13.2.0",
-    "@angular/forms": "~13.2.0",
-    "@angular/platform-browser": "~13.2.0",
-    "@angular/platform-browser-dynamic": "~13.2.0",
-    "@angular/router": "~13.2.0",
-    "@nrwl/angular": "13.8.5",
+    "@angular/animations": "13.3.11",
+    "@angular/cdk": "13.3.8",
+    "@angular/common": "13.3.11",
+    "@angular/compiler": "13.3.11",
+    "@angular/core": "13.3.11",
+    "@angular/forms": "13.3.11",
+    "@angular/platform-browser": "13.3.11",
+    "@angular/platform-browser-dynamic": "13.3.11",
+    "@angular/router": "13.3.11",
+    "@nrwl/angular": "14.1.9",
     "animate.css": "^4.1.1",
     "i18n-iso-countries": "^7.4.0",
     "normalize.css": "^8.0.1",
@@ -33,32 +34,33 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.2.0",
+    "@angular-devkit/build-angular": "13.3.7",
     "@angular-eslint/eslint-plugin": "~13.0.1",
     "@angular-eslint/eslint-plugin-template": "~13.0.1",
     "@angular-eslint/template-parser": "~13.0.1",
-    "@angular/cli": "~13.2.0",
-    "@angular/compiler-cli": "~13.2.0",
-    "@angular/language-service": "~13.2.0",
-    "@nrwl/cli": "13.8.5",
-    "@nrwl/cypress": "13.8.5",
-    "@nrwl/eslint-plugin-nx": "13.8.5",
-    "@nrwl/jest": "13.8.5",
-    "@nrwl/linter": "13.8.5",
-    "@nrwl/tao": "13.8.5",
-    "@nrwl/workspace": "13.8.5",
-    "@types/jest": "27.0.2",
+    "@angular/cli": "13.3.7",
+    "@angular/compiler-cli": "13.3.11",
+    "@angular/language-service": "13.3.11",
+    "@nrwl/cli": "14.1.9",
+    "@nrwl/cypress": "14.1.9",
+    "@nrwl/eslint-plugin-nx": "14.1.9",
+    "@nrwl/jest": "14.1.9",
+    "@nrwl/linter": "14.1.9",
+    "@nrwl/workspace": "14.1.9",
+    "@types/jest": "27.4.1",
     "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "~5.10.0",
-    "@typescript-eslint/parser": "~5.10.0",
+    "@typescript-eslint/eslint-plugin": "5.18.0",
+    "@typescript-eslint/parser": "5.18.0",
     "cypress": "^9.1.0",
-    "eslint": "~8.7.0",
+    "eslint": "8.12.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
-    "jest": "27.2.3",
+    "jest": "27.5.1",
     "jest-preset-angular": "11.1.1",
     "prettier": "^2.5.1",
-    "ts-jest": "27.0.5",
-    "typescript": "~4.5.2"
+    "ts-jest": "27.1.4",
+    "typescript": "4.6.4",
+    "nx": "14.1.9"
   }
 }
+
