@@ -1,12 +1,8 @@
-import { Component, OnInit } from '@angular/core'
+import { Component } from '@angular/core'
 
 @Component({
   selector: 'admin-shell',
   templateUrl: './shell.component.html',
-  styleUrls: ['./shell.component.scss'],
 })
-export class ShellComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit(): void {}
+export class ShellComponent {
 }
