@@ -1,12 +1,9 @@
-import { Component, OnInit } from '@angular/core'
+import { Component } from '@angular/core'
 
 @Component({
   selector: 'my-qart-footer',
   templateUrl: './footer.component.html',
-  styleUrls: ['./footer.component.scss'],
 })
-export class FooterComponent implements OnInit {
-  constructor() {}
+export class FooterComponent {
 
-  ngOnInit(): void {}
 }
