@@ -1,11 +1,8 @@
-import { Component, OnInit } from '@angular/core'
+import { Component } from '@angular/core'
 
 @Component({
   selector: 'my-qart-header',
   templateUrl: './header.component.html',
 })
-export class HeaderComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit(): void {}
+export class HeaderComponent {
 }
