@@ -1,12 +1,9 @@
-import { Component, OnInit } from '@angular/core'
+import { Component } from '@angular/core'
 
 @Component({
   selector: 'my-qart-nav',
   templateUrl: './nav.component.html',
   styles: [],
 })
-export class NavComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit(): void {}
+export class NavComponent {
 }
