@@ -1,5 +1,8 @@
 import { Component, Input, OnInit } from '@angular/core'
 import { Product } from '../../models/product'
+import {CartItem, CartService} from "@my-qart/orders";
+
+
 
 @Component({
   selector: 'products-product-item',
@@ -9,7 +12,15 @@ import { Product } from '../../models/product'
 export class ProductItemComponent implements OnInit {
   @Input() product: Product
 
-  constructor() {}
+  constructor(private cartService:CartService) {}
 
   ngOnInit(): void {}
+
+  addProductToCart() {
+    const cartItem: CartItem = {
+      productId: this.product.id,
+      quantity: 1
+    }
+    this.cartService.setCartItem(cartItem)
+  }
 }
